bpftrace: Fix cpid for Big endian
Issue:
- cpid is written to higher 32 bits in BE
- bpftrace -v -e 'BEGIN { printf("%d", cpid); }'
Bytecode:
0: (bf) r6 = r1
1: (b7) r1 = 0
2: (7b) *(u64 *)(r10 -8) = r1
last_idx 2 first_idx 0
regs=2 stack=0 before 1: (b7) r1 = 0
3: (b7) r2 = 122768
4: (63) *(u32 *)(r10 -8) = r2 <-----
5: (7b) *(u64 *)(r10 -16) = r1
(...)

Fix:
- Promote it to 64 bits.

+++ src/ast/codegen_llvm.cpp
@@ -270,7 +270,7 @@ void CodegenLLVM::visit(Builtin &builtin)
       std::cerr << "BUG: Invalid cpid: " << cpid << std::endl;
       abort();
     }
-    expr_ = b_.getInt32(cpid);
+    expr_ = b_.getInt64(cpid);
   }
   else
   {
